+++ RELEASE_NOTES.md
@@ -1,12 +1,12 @@
 ## VPinFE Release Notes
 
 ### Summary
-This is the offical release of vpinfe v2.1.2
+This is the offical release of vpinfe v2.1.3
 
 ### What's New
 None
 
 ### Fixes
-- fix vpxparser to ignore commented out cgamenames
+- Fix rom identification of score parsing to VPXFile one instead of VPSDB one.
 
 ### Notes